Enhance bind-parse-type-from-method-param to comprehensive baseline test
