+++ src/libraries/Microsoft.Extensions.Configuration.Binder/tests/SourceGenerationTests/GeneratorTests.Baselines.cs
@@ -118,6 +118,49 @@ namespace Microsoft.Extensions.SourceGeneration.Configuration.Binder.Tests
         }
 
         [Fact]
+        public async Task Bind_CanParseTargetConfigType_FromMethodParam()
+        {
+            string source = """
+                using System;
+                using Microsoft.Extensions.Configuration;
+
+                public class Program
+                {
+                    public static void Main()
+                    {
+                        ConfigurationBuilder configurationBuilder = new();
+                        IConfiguration config = configurationBuilder.Build();
+
+                        BindOptions(config, new MyClass0());
+                        BindOptions(config, new MyClass1(), _ => { });
+                        BindOptions(config, "", new MyClass2());
+                    }
+
+                    private static void BindOptions(IConfiguration config, MyClass0 instance)
+                    {
+                        config.Bind(instance);
+                    }
+
+                    private static void BindOptions(IConfiguration config, MyClass1 instance, Action<BinderOptions>? configureOptions)
+                    {
+                        config.Bind(instance, configureOptions);
+                    }
+
+                    private static void BindOptions(IConfiguration config, string path, MyClass2 instance)
+                    {
+                        config.Bind(path, instance);
+                    }
+
+                    public class MyClass0 { }
+                    public class MyClass1 { }
+                    public class MyClass2 { }
+                }
+                """;
+
+            await VerifyAgainstBaselineUsingFile("Bind_ParseTypeFromMethodParam.generated.txt", source, extType: ExtensionClassType.ConfigurationBinder);
+        }
+
+        [Fact]
         public async Task Get()
         {
             string source = @"
